Annotations and memoization

+++ test/org/carbon/Helpers.kt
@@ -3,5 +3,5 @@ package org.carbon
 import org.carbon.runtime.CarbonInteger
 import org.carbon.runtime.CarbonString
 
-fun wrap(int: Int) = CarbonInteger(int)
+fun wrap(int: Long) = CarbonInteger(int)
 fun wrap(string: String) = CarbonString(string)
